Revert "Move collectibles module into collection.go (#821)"
This reverts commit 3590a80e717417721e09943ae597ab88f5332ea2.
